Cluster-autoscaler Move scale up code to separate file
